+++ src/app/react/src/components/providers/SelectorModeProvider/index.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
-import type { SelectorModeProviderProps } from '../../types';
-import { Provider } from '../../contexts/SelectorModeContext';
-import { SelectorModes } from '../../utils/selector-modes';
+import type { SelectorModeProviderProps } from '../../../types';
+import { Provider } from '../../../contexts/SelectorModeContext';
+import { SelectorModes } from '../../../utils/selector-modes';
 
 export default function SelectorModeProvider({ initialMode, children }) {
   const [mode, setMode] = useState(initialMode ?? SelectorModes.FreeMove);
